requirements.txt
# Django Core
Django>=5.0
djangorestframework>=3.14

# Authentication
djangorestframework-simplejwt>=5.3

# CORS (for React frontend)
django-cors-headers>=4.3

# Environment variables
python-dotenv>=1.0

# Image handling
Pillow>=10.0

# PostgreSQL (uncomment for production)
# psycopg2-binary>=2.9

# Development tools
# django-debug-toolbar>=4.2

# Testing
# pytest-django>=4.7
# factory-boy>=3.3
